mover thermal_speed (functionality) to plasmapy.formulary.speeds
